feat: [WIP] vehicle and travels resource

+++ src/main/java/iff/poo/core/BaseRepository.java
@@ -3,7 +3,7 @@ package iff.poo.core;
 import java.util.List;
 
 public interface BaseRepository<T> {
-    public T getById(int id);
+    public T getById(Long id);
     public Long create(T object);
     public List<T> getAll();
     public boolean updateById(T object, Long id);
